WIP: Obs timedeltas

+++ training/src/anemoi/training/data/data_reader.py
@@ -565,9 +565,9 @@ class ObservationDataReader(BaseAnemoiReader):
 
         # the leading time axis is intentionally absent — per-time
         # structure is recoverable through ``boundaries``.
-        data = torch.from_numpy(np.asarray(x.data))
-        latitudes = np.deg2rad(np.asarray(x.latitudes))
-        longitudes = np.deg2rad(np.asarray(x.longitudes))
+        data = torch.from_numpy(np.asarray(x.data, dtype=np.float32))
+        latitudes = np.deg2rad(np.asarray(x.latitudes, dtype=np.float32))
+        longitudes = np.deg2rad(np.asarray(x.longitudes, dtype=np.float32))
         coordinates = torch.from_numpy(
             np.stack([latitudes, longitudes], axis=-1),
         )
